Moved configurationSourceTracker to a sub folder
